pyproject.toml
[tool.poetry]
name = "geminipoc"
version = "0.1.0"
description = ""
authors = ["Advika101 <hannah13200@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-generativeai = "^0.6.0"
google-api-python-client = "^2.132.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"
pdf2image = "^1.17.0"
python-dotenv = "^1.0.1"
ipython = "^8.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
